src/modules/rentals/useCases/devolutionRental/DevolutionRentalController.ts
import { Request, Response } from 'express';
import { container } from 'tsyringe';

import { DevolutionRentalUseCase } from './DevolutionRentalUseCase';

class DevolutionRentalController {
  async handle(request: Request, response: Response): Promise<Response> {
    const { rental_id } = request.params;

    const devolutionRentalUseCase = container.resolve(DevolutionRentalUseCase);

    const rental = await devolutionRentalUseCase.execute({ id: rental_id });

    return response.status(200).json(rental);
  }
}

export { DevolutionRentalController };
